current tests

+++ mesh_edit/src/marchingcubes.cpp
@@ -328,16 +328,16 @@ int IndicatorFunction(CGL::Vector3D point, OctreeNode currentNode)
 {
     //Get average point and average node from currentNode
     Vector3D projectedPoint=currentNode.projectPoint(point);
-    Vector3D direction = projectedPoint-point;
+    Vector3D direction= projectedPoint-point;
     cout<<"Original Direction: ";
     cout<<direction<<endl;
-    float dir_magnitude = sqrt(pow(direction.x,2)+pow(direction.y,2)+pow(direction.z,2));
+    double dir_magnitude = sqrt(pow(direction.x,2)+pow(direction.y,2)+pow(direction.z,2));
     if(dir_magnitude==0){
         return 0;
     }
     direction = direction/dir_magnitude;
 
-    float normal_magnitude = sqrt(pow(currentNode.avgNorm.x,2)+pow(currentNode.avgNorm.y,2)+pow(currentNode.avgNorm.z,2));
+    double normal_magnitude = sqrt(pow(currentNode.avgNorm.x, 2)+pow(currentNode.avgNorm.y, 2)+pow(currentNode.avgNorm.z, 2));
     Vector3D unit_normal = currentNode.avgNorm/normal_magnitude;
 
     cout<<"Original Corner Point: ";
@@ -350,6 +350,7 @@ int IndicatorFunction(CGL::Vector3D point, OctreeNode currentNode)
     cout<<direction<<endl;
     cout<<"Unit Normal: ";
     cout<<unit_normal<<endl;
+
     bool within_x = direction.x<=unit_normal.x+0.01 && direction.x>=unit_normal.x-0.01;
     bool within_y = direction.y<=unit_normal.y+0.01 && direction.y>=unit_normal.y-0.01;
     bool within_z = direction.z<=unit_normal.z+0.01 && direction.z>=unit_normal.z-0.01;
